test: fix naming warnings in tests
- Rename test functions to snake_case:
  - test_abstractPropertyNegative_errors -> test_abstract_property_negative_errors
  - test_async_stream_pattern_getReader -> test_async_stream_pattern_get_reader
  - test_no_predicate_NOT_compatible_with_type_guard -> test_no_predicate_not_compatible_with_type_guard
  - test_type_guard_NOT_compatible_with_asserts -> test_type_guard_not_compatible_with_asserts

- Rename test variables to snake_case (avoid conflicts):
  - lit_A, lit_B -> lit_a_upper, lit_b_upper
  - lit_ABC, lit_DEF -> lit_abc_upper, lit_def_upper
  - lit_Name, lit_Value -> lit_name_upper, lit_value_upper

- Restore type imports in solver test files that were removed by auto-fix

Final status:
- 0 clippy warnings (lib)
- 10 clippy warnings (tests, excluding "file is loaded as a module multiple times")
- 0 cargo build warnings
- All 10,110 tests passing

+++ src/cli/reporter.rs
@@ -107,13 +107,13 @@ impl Reporter {
         let mut underline = String::new();
         for (i, ch) in line_text.chars().enumerate() {
             let offset = i as u32;
-            if offset < column as u32 - 1 {
+            if offset < column - 1 {
                 if ch == '\t' {
                     underline.push_str("    "); // Tabs expand to 4 spaces
                 } else {
                     underline.push(' ');
                 }
-            } else if offset >= column as u32 - 1 && offset < end {
+            } else if offset >= column - 1 && offset < end {
                 // Within the error span - use tildes (or ~ for single char)
                 if ch == '\t' {
                     underline.push_str("~~~~");
